Added floating point.
I need to do more research here.  I *think* all 32-bit integer maths
is exact in doubles, but not 100% sure.  If so the DBL_EPSILON tweaks
aren't needed.

Also this isn't complete as <= doesn't have the same logic either.
Hence for now we're not doing "close-enough" checks.

+++ expr.h
@@ -25,7 +25,7 @@ DEALINGS IN THE SOFTWARE.  */
 #ifndef EXPR_H
 #define EXPR_H
 
-typedef int (sym_func)(void *data, char *str, char **end, int *err);
-int evaluate_filter(void *data, sym_func *f, char *str, int *err);
+typedef double (sym_func)(void *data, char *str, char **end, int *err);
+double evaluate_filter(void *data, sym_func *f, char *str, int *err);
 
 #endif /* EXPR_H */
